Refactoring

+++ src/main/java/com/crmw/CRM_BE/entity/Users.java
@@ -1,7 +1,6 @@
 package com.crmw.CRM_BE.entity;
 
 import jakarta.persistence.*;
-import org.hibernate.boot.model.internal.XMLContext;
 
 import java.time.Instant;
 
@@ -11,7 +10,7 @@ public class Users {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(unique = true, nullable = false)
     private String username;
@@ -36,7 +35,7 @@ public class Users {
         this.createdAt = createdAt;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -52,7 +51,7 @@ public class Users {
         this.role = role;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
